+++ src/DpressException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Dynart\Dpress;
+
+use Dynart\Micro\MicroException;
+
+class DpressException extends MicroException {}